requirements.txt
flask
tensorflow
opencv-python
numpy
matplotlib
seaborn
Pillow
mysql-connector-python
